Refactor and cleanup routing and models
